update font-face mixin to allow weight and style to be set properly
Essentially it would appear that we've been doing this the hard way for years.
It's possible to set font-weight and font-style in the font-face mixin as long as
it's done in a specific order. Font-family must be first, then src, the font-weight
and font-style can be set.
This will mean we no longer need to use the `font` mixin to set fonts each time
we need to change the weight or style. We can simply use `font-family` to set
the web font family and then change the weight and style as needed on child elements
with font-weight and font-style